"clean code of run + create a file utils.py to put all minor methods"
